pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "program-translator"
description = "专业文档翻译工具，通过专有名词识别和正则过滤确保术语准确翻译"
authors = [{name = "User"}]
requires-python = ">=3.11"
version = "1.1.0"

dependencies = [
    "fastapi==0.115.12",
    "uvicorn==0.34.2",
    "pydantic==2.11.9",
    "pydantic-settings==2.1.0",
    "python-multipart==0.0.20",
    "pandas==2.3.2",
    "requests==2.32.5",
    "numpy==2.3.3",
    "python-dotenv==1.1.0",
    "markdown==3.8",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "azure-ai-documentintelligence==1.0.2",
    "azure-core==1.36.0",
    "azure-identity==1.25.1",
    "beautifulsoup4==4.13.4",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "cryptography==46.0.3",
    "distro==1.9.0",
    "dotenv==0.9.9",
    "filelock==3.19.1",
    "flatbuffers==25.2.10",
    "fsspec==2025.9.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==0.35.0",
    "humanfriendly==10.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "isodate==0.7.2",
    "jinja2==3.1.6",
    "jiter==0.11.0",
    "joblib==1.5.2",
    "magika==0.6.2",
    "markdownify==1.2.0",
    "markitdown[docx,pdf,pptx]==0.1.1",
    "markupsafe==3.0.2",
    "mpmath==1.3.0",
    "msal==1.34.0",
    "msal-extensions==1.3.1",
    "networkx==3.5",
    "nltk==3.9.1",
    "onnxruntime==1.20.1",
    "openai==1.82.0",
    "packaging==25.0",
    "pdfminer-six==20251107",
    "pip==25.2",
    "pluggy==1.6.0",
    "protobuf==6.32.1",
    "pycparser==2.23",
    "pydantic-core==2.33.2",
    "pyjwt==2.10.1",
    "pyreadline3==3.5.4",
    "pytest==8.3.5",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "regex==2025.9.1",
    "safetensors==0.6.2",
    "setuptools==65.5.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "starlette==0.46.2",
    "sympy==1.14.0",
    "tokenizers==0.22.0",
    "torch==2.8.0",
    "tqdm==4.67.1",
    "transformers==4.56.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "google-genai>=1.52.0",
]

[tool.uv]
# Python版本要求已在project.required-python中设置

[tool.setuptools]
packages = ["modules", "services"]

[project.scripts]
program-translator-cli = "main:main"
program-translator-webui = "app:main"
